Refactored directory structure and "requires"
Because constants.rb is now dependant on the unicode char ranges (at
compilation time, not just run time), the files must be required in a
more well-defined order.

Also, the core extensions folder was in a bit of a weird palce, now
moved.
