+++ lib/core_extensions/regexp/examples.rb
